+++ compiler/smScSema.c
@@ -277,7 +277,7 @@ int smScAction(
 	do {
 	  i = smScSkip(smThis, 1);
 	  c = smThis->smText[smThis->smLength-1];
-	} while (isspace(c) && i > 0);
+	} while (isspace((int)c) && i > 0);
 
 	if (c != '.') {
 	  lmLog(&token.srcp, 109, sevERR, "expected terminating '.'");
